host/pyproject.toml
[project]
name = "pideck"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.6",
    "pillow>=11.2.1",
    "pyserial>=3.5",
]
